modified:   .htaccess 	modified:   ajax/theme.ajax.php 	modified:   ajax/users.ajax.php 	modified:   controllers/categories.controller.php 	modified:   controllers/customers.controller.php 	modified:   controllers/products.controller.php 	modified:   controllers/users.controller.php 	new file:   helpers/security.php 	modified:   index.php 	modified:   models/users.model.php 	modified:   views/js/categories.js 	modified:   views/js/customers.js 	modified:   views/js/products.js 	modified:   views/js/template.js 	modified:   views/js/users.js 	modified:   views/modules/categories.php 	modified:   views/modules/customers.php 	modified:   views/modules/header.php 	modified:   views/modules/products.php 	modified:   views/modules/users.php 	modified:   views/template.php

+++ controllers/users.controller.php
@@ -5,211 +5,169 @@ class ControllerUsers{
 	/*=============================================
 	USER LOGIN
 	=============================================*/
-	
+
 	static public function ctrUserLogin(){
 
 		if (isset($_POST["loginUser"])) {
-			
-			if (preg_match('/^[a-zA-Z0-9]+$/', $_POST["loginUser"]) && 
-				preg_match('/^[a-zA-Z0-9]+$/', $_POST["loginPass"])) {
 
-				$encryptpass = crypt($_POST["loginPass"], '$2a$07$asxx54ahjppf45sd87a5a4dDDGsystemdev$');
-				
+			$loginUser = $_POST["loginUser"];
+			$loginPass = $_POST["loginPass"];
+
+			if (preg_match('/^[a-zA-Z0-9]+$/', $loginUser) &&
+				strlen($loginPass) > 0 && strlen($loginPass) <= 72) {
+
 				$table = 'users';
+				$item  = 'user';
+				$answer = UsersModel::MdlShowUsers($table, $item, $loginUser);
 
-				$item = 'user';
-				$value = $_POST["loginUser"];
+				if ($answer && $answer["user"] === $loginUser) {
 
-				$answer = UsersModel::MdlShowUsers($table, $item, $value);
+					$verified = false;
 
-				// var_dump($answer);
-
-				if($answer["user"] == $_POST["loginUser"] && $answer["password"] == $encryptpass){
-
-					if($answer["status"] == 1){
-
-						$_SESSION["loggedIn"] = "ok";
-						$_SESSION["id"] = $answer["id"];
-						$_SESSION["name"] = $answer["name"];
-						$_SESSION["user"] = $answer["user"];
-						$_SESSION["photo"] = $answer["photo"];
-						$_SESSION["profile"] = $answer["profile"];
-
-						/*=============================================
-						Register date to know last_login
-						=============================================*/
-
-						date_default_timezone_set("America/Bogota");
-
-						$date = date('Y-m-d');
-						$hour = date('H:i:s');
-
-						$actualDate = $date.' '.$hour;
-
-						$item1 = "lastLogin";
-						$value1 = $actualDate;
-
-						$item2 = "id";
-						$value2 = $answer["id"];
-
-						$lastLogin = UsersModel::mdlUpdateUser($table, $item1, $value1, $item2, $value2);
-
-						if($lastLogin == "ok"){
-
-							echo '<script>
-
-								window.location = "home";
-
-							</script>';
-
-						}
-
-					}else{
-						
-						echo '<br><div class="alert alert-danger">User is deactivated</div>';
-					
+					// Try modern password_hash format first
+					if (password_verify($loginPass, $answer["password"])) {
+						$verified = true;
+					}
+					// Fall back to legacy crypt and rehash on success
+					elseif ($answer["password"] === crypt($loginPass, '$2a$07$asxx54ahjppf45sd87a5a4dDDGsystemdev$')) {
+						$verified = true;
+						$newHash = password_hash($loginPass, PASSWORD_DEFAULT);
+						UsersModel::mdlUpdateUser($table, 'password', $newHash, 'id', $answer["id"]);
 					}
 
-				}else{
+					if ($verified) {
 
+						if ($answer["status"] == 1) {
+
+							$_SESSION["loggedIn"] = "ok";
+							$_SESSION["id"]      = $answer["id"];
+							$_SESSION["name"]    = $answer["name"];
+							$_SESSION["user"]    = $answer["user"];
+							$_SESSION["photo"]   = $answer["photo"];
+							$_SESSION["profile"] = $answer["profile"];
+
+							date_default_timezone_set("America/Bogota");
+							$actualDate = date('Y-m-d H:i:s');
+
+							$lastLogin = UsersModel::mdlUpdateUser($table, 'lastLogin', $actualDate, 'id', $answer["id"]);
+
+							if ($lastLogin == "ok") {
+								echo '<script>window.location = "home";</script>';
+							}
+
+						} else {
+							echo '<br><div class="alert alert-danger">User is deactivated</div>';
+						}
+
+					} else {
+						echo '<br><div class="alert alert-danger">User or password incorrect</div>';
+					}
+
+				} else {
 					echo '<br><div class="alert alert-danger">User or password incorrect</div>';
-				
 				}
-			
+
 			}
-		
+
 		}
-	
+
 	}
 
 
 	/*=============================================
 	CREATE USER
 	=============================================*/
-	
+
 	static public function ctrCreateUser(){
 
 		if (isset($_POST["newUser"])) {
-			
+
+			csrf_verify();
+
 			if (preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["newName"]) &&
 				preg_match('/^[a-zA-Z0-9]+$/', $_POST["newUser"]) &&
-				preg_match('/^[a-zA-Z0-9]+$/', $_POST["newPasswd"])){
-
-				/*=============================================
-				VALIDATE IMAGE
-				=============================================*/
+				strlen($_POST["newPasswd"]) > 0) {
 
 				$photo = "";
-			
-				if (isset($_FILES["newPhoto"]["tmp_name"])){
+
+				if (isset($_FILES["newPhoto"]["tmp_name"]) && !empty($_FILES["newPhoto"]["tmp_name"])) {
 
 					list($width, $height) = getimagesize($_FILES["newPhoto"]["tmp_name"]);
-					
-					$newWidth = 500;
+
+					$newWidth  = 500;
 					$newHeight = 500;
 
-					/*=============================================
-					Let's create the folder for each user
-					=============================================*/
-
-					$folder = "views/img/users/".$_POST["newUser"];
+					// Strip everything except alphanumeric to prevent path traversal
+					$safeUser = preg_replace('/[^a-zA-Z0-9]/', '', $_POST["newUser"]);
+					$folder   = "views/img/users/" . $safeUser;
 
 					mkdir($folder, 0755);
 
-					/*=============================================
-					PHP functions depending on the image
-					=============================================*/
+					if ($_FILES["newPhoto"]["type"] == "image/jpeg") {
 
-					if($_FILES["newPhoto"]["type"] == "image/jpeg"){
-
-						$randomNumber = mt_rand(100,999);
-						
-						$photo = "views/img/users/".$_POST["newUser"]."/".$randomNumber.".jpg";
-						
+						$randomNumber = mt_rand(100, 999);
+						$photo = "views/img/users/" . $safeUser . "/" . $randomNumber . ".jpg";
 						$srcImage = imagecreatefromjpeg($_FILES["newPhoto"]["tmp_name"]);
-						
 						$destination = imagecreatetruecolor($newWidth, $newHeight);
-
 						imagecopyresized($destination, $srcImage, 0, 0, 0, 0, $newWidth, $newHeight, $width, $height);
-
 						imagejpeg($destination, $photo);
 
 					}
 
 					if ($_FILES["newPhoto"]["type"] == "image/png") {
 
-						$randomNumber = mt_rand(100,999);
-						
-						$photo = "views/img/users/".$_POST["newUser"]."/".$randomNumber.".png";
-						
+						$randomNumber = mt_rand(100, 999);
+						$photo = "views/img/users/" . $safeUser . "/" . $randomNumber . ".png";
 						$srcImage = imagecreatefrompng($_FILES["newPhoto"]["tmp_name"]);
-						
 						$destination = imagecreatetruecolor($newWidth, $newHeight);
-
 						imagecopyresized($destination, $srcImage, 0, 0, 0, 0, $newWidth, $newHeight, $width, $height);
-
 						imagepng($destination, $photo);
 					}
 
 				}
 
-				$table = 'users';
+				$table      = 'users';
+				$encryptpass = password_hash($_POST["newPasswd"], PASSWORD_DEFAULT);
 
-				$encryptpass = crypt($_POST["newPasswd"], '$2a$07$asxx54ahjppf45sd87a5a4dDDGsystemdev$');
-
-				$data = array('name' => $_POST["newName"],
-							  'user' => $_POST["newUser"],
-								'password' => $encryptpass,
-								'profile' => $_POST["newProfile"],
-								'photo' => $photo);
+				$data = array(
+					'name'     => $_POST["newName"],
+					'user'     => $_POST["newUser"],
+					'password' => $encryptpass,
+					'profile'  => $_POST["newProfile"],
+					'photo'    => $photo,
+				);
 
 				$answer = UsersModel::mdlAddUser($table, $data);
 
 				if ($answer == 'ok') {
 
-						echo '<script>
-						
-						swal({
-							type: "success",
-							title: "User added succesfully!",
-							showConfirmButton: true,
-							confirmButtonText: "Close"
-
-						}).then(function(result){
-
-							if(result.value){
-
-								window.location = "users";
-							}
-
-						});
-						
-						</script>';
+					echo '<script>
+					swal({
+						type: "success",
+						title: "User added succesfully!",
+						showConfirmButton: true,
+						confirmButtonText: "Close"
+					}).then(function(result){
+						if (result.value) { window.location = "users"; }
+					});
+					</script>';
 
 				}
-			
-			}else{
+
+			} else {
 
 				echo '<script>
-					
 					swal({
 						type: "error",
 						title: "No special characters or blank fields",
 						showConfirmButton: true,
 						confirmButtonText: "Close"
-			
-						}).then(function(result){
-
-							if(result.value){
-
-								window.location = "users";
-							}
-
-						});
-					
+					}).then(function(result){
+						if (result.value) { window.location = "users"; }
+					});
 				</script>';
 			}
-			
+
 		}
 	}
 
@@ -219,10 +177,8 @@ class ControllerUsers{
 
 	static public function ctrShowUsers($item, $value){
 
-		$table = "users";
-
+		$table  = "users";
 		$answer = UsersModel::MdlShowUsers($table, $item, $value);
-
 		return $answer;
 	}
 
@@ -233,176 +189,106 @@ class ControllerUsers{
 	static public function ctrEditUser(){
 
 		if (isset($_POST["EditUser"])) {
-			
-			if (preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["EditName"])){
 
-				/*=============================================
-				VALIDATE IMAGE
-				=============================================*/
+			csrf_verify();
+
+			if (preg_match('/^[a-zA-Z0-9ñÑáéíóúÁÉÍÓÚ ]+$/', $_POST["EditName"])) {
 
 				$photo = $_POST["currentPicture"];
 
-				if(isset($_FILES["editPhoto"]["tmp_name"]) && !empty($_FILES["editPhoto"]["tmp_name"])){
+				if (isset($_FILES["editPhoto"]["tmp_name"]) && !empty($_FILES["editPhoto"]["tmp_name"])) {
 
 					list($width, $height) = getimagesize($_FILES["editPhoto"]["tmp_name"]);
-					
-					$newWidth = 500;
+
+					$newWidth  = 500;
 					$newHeight = 500;
 
-					/*=============================================
-					Let's create the folder for each user
-					=============================================*/
+					$safeUser = preg_replace('/[^a-zA-Z0-9]/', '', $_POST["EditUser"]);
+					$folder   = "views/img/users/" . $safeUser;
 
-					$folder = "views/img/users/".$_POST["EditUser"];
-
-					/*=============================================
-					we ask first if there's an existing image in the database
-					=============================================*/
-
-					if (!empty($_POST["currentPicture"])){
-						
+					if (!empty($_POST["currentPicture"]) &&
+						strpos($_POST["currentPicture"], 'views/img/users/') === 0) {
 						unlink($_POST["currentPicture"]);
-
-					}else{
-
+					} else {
 						mkdir($folder, 0755);
-
 					}
 
-					/*=============================================
-					PHP functions depending on the image
-					=============================================*/
+					if ($_FILES["editPhoto"]["type"] == "image/jpeg") {
 
-					if($_FILES["editPhoto"]["type"] == "image/jpeg"){
-
-						/*We save the image in the folder*/
-
-						$randomNumber = mt_rand(100,999);
-						
-						$photo = "views/img/users/".$_POST["EditUser"]."/".$randomNumber.".jpg";
-						
+						$randomNumber = mt_rand(100, 999);
+						$photo = "views/img/users/" . $safeUser . "/" . $randomNumber . ".jpg";
 						$srcImage = imagecreatefromjpeg($_FILES["editPhoto"]["tmp_name"]);
-						
 						$destination = imagecreatetruecolor($newWidth, $newHeight);
-
 						imagecopyresized($destination, $srcImage, 0, 0, 0, 0, $newWidth, $newHeight, $width, $height);
-
 						imagejpeg($destination, $photo);
 
 					}
-					
+
 					if ($_FILES["editPhoto"]["type"] == "image/png") {
 
-						/*We save the image in the folder*/
-
-						$randomNumber = mt_rand(100,999);
-						
-						$photo = "views/img/users/".$_POST["EditUser"]."/".$randomNumber.".png";
-						
+						$randomNumber = mt_rand(100, 999);
+						$photo = "views/img/users/" . $safeUser . "/" . $randomNumber . ".png";
 						$srcImage = imagecreatefrompng($_FILES["editPhoto"]["tmp_name"]);
-						
 						$destination = imagecreatetruecolor($newWidth, $newHeight);
-
 						imagecopyresized($destination, $srcImage, 0, 0, 0, 0, $newWidth, $newHeight, $width, $height);
-
 						imagepng($destination, $photo);
 					}
 
 				}
 
-				
 				$table = 'users';
 
-				if($_POST["EditPasswd"] != ""){
+				if ($_POST["EditPasswd"] != "") {
 
-					if(preg_match('/^[a-zA-Z0-9]+$/', $_POST["EditPasswd"])){
+					$encryptpass = password_hash($_POST["EditPasswd"], PASSWORD_DEFAULT);
 
-						$encryptpass = crypt($_POST["EditPasswd"], '$2a$07$asxx54ahjppf45sd87a5a4dDDGsystemdev$');
-
-					}
-
-					else{
-
-						echo '<script>
-					
-							swal({
-								type: "error",
-								title: "No special characters in the password or blank fields",
-								showConfirmButton: true,
-								confirmButtonText: "Close"
-
-								}).then(function(result){
-										
-									if (result.value) {
-						
-										window.location = "users";
-
-									}
-								});
-							
-						</script>';
-					}
-				
-				}else{
+				} else {
 
 					$encryptpass = $_POST["currentPasswd"];
-					
+
 				}
 
-				$data = array('name' => $_POST["EditName"],
-								'user' => $_POST["EditUser"],
-								'password' => $encryptpass,
-								'profile' => $_POST["EditProfile"],
-								'photo' => $photo);
+				$data = array(
+					'name'     => $_POST["EditName"],
+					'user'     => $_POST["EditUser"],
+					'password' => $encryptpass,
+					'profile'  => $_POST["EditProfile"],
+					'photo'    => $photo,
+				);
 
 				$answer = UsersModel::mdlEditUser($table, $data);
 
 				if ($answer == 'ok') {
-					
+
 					echo '<script>
-					
-						swal({
-							type: "success",
-							title: "User edited succesfully!",
-							showConfirmButton: true,
-							confirmButtonText: "Close"
+					swal({
+						type: "success",
+						title: "User edited succesfully!",
+						showConfirmButton: true,
+						confirmButtonText: "Close"
+					}).then(function(result){
+						if (result.value) { window.location = "users"; }
+					});
+					</script>';
 
-						 }).then(function(result){
-							
-							if (result.value) {
+				} else {
 
-								window.location = "users";
-							}
-
-						});
-					
+					echo '<script>
+					swal({
+						type: "error",
+						title: "No special characters in the name or blank field",
+						showConfirmButton: true,
+						confirmButtonText: "Close"
+					}).then(function(result){
+						if (result.value) { window.location = "users"; }
+					});
 					</script>';
 				}
-				else{
-					echo '<script>
-						
-						swal({
-							type: "error",
-							title: "No special characters in the name or blank field",
-							showConfirmButton: true,
-							confirmButtonText: "Close"
-							 }).then(function(result){
-									
-								if (result.value) {
 
-									window.location = "users";
-								
-								}
+			}
 
-							});
-						
-					</script>';
-				}
-			
-			}	
-		
 		}
-	
+
 	}
 
 	/*=============================================
@@ -411,46 +297,44 @@ class ControllerUsers{
 
 	static public function ctrDeleteUser(){
 
-		if(isset($_GET["userId"])){
+		if (isset($_GET["userId"])) {
 
-			$table ="users";
-			$data = $_GET["userId"];
+			csrf_verify();
 
-			if($_GET["userPhoto"] != ""){
+			$table = "users";
+			$data  = $_GET["userId"];
 
-				unlink($_GET["userPhoto"]);				
-				rmdir('views/img/users/'.$_GET["username"]);
+			// Validate paths before filesystem operations
+			if (!empty($_GET["userPhoto"]) &&
+				strpos($_GET["userPhoto"], 'views/img/users/') === 0 &&
+				file_exists($_GET["userPhoto"])) {
+				unlink($_GET["userPhoto"]);
+			}
 
+			if (!empty($_GET["username"]) &&
+				preg_match('/^[a-zA-Z0-9]+$/', $_GET["username"])) {
+				@rmdir('views/img/users/' . $_GET["username"]);
 			}
 
 			$answer = UsersModel::mdlDeleteUser($table, $data);
 
-			if($answer == "ok"){
+			if ($answer == "ok") {
 
-				echo'<script>
-
+				echo '<script>
 				swal({
-					  type: "success",
-					  title: "The user has been succesfully deleted",
-					  showConfirmButton: true,
-					  confirmButtonText: "Close"
-
-					  }).then(function(result){
-					  	
-						if (result.value) {
-
-						window.location = "users";
-
-						}
-					})
-
+					type: "success",
+					title: "The user has been succesfully deleted",
+					showConfirmButton: true,
+					confirmButtonText: "Close"
+				}).then(function(result){
+					if (result.value) { window.location = "users"; }
+				})
 				</script>';
 
-			}		
+			}
 
 		}
 
 	}
-	
-}
 
+}
